# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/material-automation/status": {
      "get": {
        "operationId": "MaterialsController_automationStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/material-automation/scan": {
      "post": {
        "operationId": "MaterialsController_runAutomationScan",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/material-types": {
      "get": {
        "operationId": "MaterialsController_types",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/students/{studentId}/materials": {
      "get": {
        "operationId": "MaterialsController_list",
        "parameters": [
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      },
      "post": {
        "operationId": "MaterialsController_create",
        "parameters": [
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMaterialItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/materials/{materialId}/versions": {
      "post": {
        "operationId": "MaterialsController_upload",
        "parameters": [
          {
            "name": "materialId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadMaterialVersionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/materials/{materialId}/review": {
      "post": {
        "operationId": "MaterialsController_review",
        "parameters": [
          {
            "name": "materialId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewMaterialDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/materials/{materialId}/mark-missing": {
      "post": {
        "operationId": "MaterialsController_markMissing",
        "parameters": [
          {
            "name": "materialId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkMaterialMissingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/materials/{materialId}/followups": {
      "post": {
        "operationId": "MaterialsController_followup",
        "parameters": [
          {
            "name": "materialId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialFollowupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/materials/{materialId}/archive": {
      "post": {
        "operationId": "MaterialsController_archive",
        "parameters": [
          {
            "name": "materialId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArchiveMaterialDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/materials/versions/{versionId}/download": {
      "get": {
        "operationId": "MaterialsController_download",
        "parameters": [
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/material-submissions/{submissionId}": {
      "get": {
        "operationId": "MaterialsController_submissionDetail",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/materials/{materialId}/submissions": {
      "post": {
        "operationId": "MaterialsController_createSubmission",
        "parameters": [
          {
            "name": "materialId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMaterialSubmissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/material-submissions/{submissionId}/files": {
      "post": {
        "operationId": "MaterialsController_addSubmissionFile",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadMaterialVersionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/material-submissions/{submissionId}/files/{fileId}/remove": {
      "post": {
        "operationId": "MaterialsController_removeSubmissionFile",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveMaterialSubmissionFileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/material-submissions/{submissionId}/submit": {
      "post": {
        "operationId": "MaterialsController_submitSubmission",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/material-submissions/{submissionId}/withdraw": {
      "post": {
        "operationId": "MaterialsController_withdrawSubmission",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawMaterialSubmissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/material-submissions/{submissionId}/review/start": {
      "post": {
        "operationId": "MaterialsController_startSubmissionReview",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/material-submissions/{submissionId}/review": {
      "post": {
        "operationId": "MaterialsController_reviewSubmission",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewMaterialSubmissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/material-applicability-requests/{requestId}/review": {
      "post": {
        "operationId": "MaterialsController_reviewApplicability",
        "parameters": [
          {
            "name": "requestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewMaterialApplicabilityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/materials/{materialId}/cancel": {
      "post": {
        "operationId": "MaterialsController_cancelSpecial",
        "parameters": [
          {
            "name": "materialId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelSpecialMaterialDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/material-submission-files/{fileId}/download": {
      "get": {
        "operationId": "MaterialsController_downloadSubmissionFile",
        "parameters": [
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "materials"
        ]
      }
    },
...
  },
  "info": {
    "title": "DSE升学服务CRM API",
    "description": "S1认证、权限、账号、审计与任务监督闭环接口",
    "version": "0.2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dse_session"
      }
    },
    "schemas": {
...
      "CreateMaterialItemDto": {
        "type": "object",
        "properties": {
          "materialTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "maxLength": 150
          },
          "requirement": {
            "type": "object",
            "maxLength": 1000,
            "nullable": true
          },
          "dueAt": {
            "type": "object",
            "format": "date-time",
            "nullable": true
          },
          "ownerId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "creationReason": {
            "type": "string",
            "maxLength": 1000,
            "description": "新增特殊资料的业务原因"
          },
          "requirementKind": {
            "type": "string",
            "enum": [
              "REQUIRED",
              "OPTIONAL"
            ],
            "default": "REQUIRED"
          }
        },
        "required": [
          "materialTypeId",
          "title",
          "creationReason"
        ]
      },
      "UploadMaterialVersionDto": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "maxLength": 255
          },
          "mimeType": {
            "type": "string",
            "maxLength": 100
          },
          "contentBase64": {
            "type": "string",
            "description": "Base64编码文件内容，解码后最大50MB"
          },
          "replacesFileId": {
            "type": "object",
            "format": "uuid",
            "nullable": true,
            "description": "补正时被替换的问题文件"
          }
        },
        "required": [
          "fileName",
          "mimeType",
          "contentBase64"
        ]
      },
      "ReviewMaterialDto": {
        "type": "object",
        "properties": {
          "outcome": {
            "type": "string",
            "enum": [
              "APPROVED",
              "PARTIALLY_MISSING",
              "RESUBMISSION_REQUIRED",
              "AWAITING_CONFIRMATION",
              "NOT_APPLICABLE"
            ]
          },
          "comment": {
            "type": "object",
            "maxLength": 1000,
            "nullable": true
          },
          "version": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "outcome",
          "version"
        ]
      },
      "MarkMaterialMissingDto": {
        "type": "object",
        "properties": {
          "missingReason": {
            "type": "string",
            "maxLength": 1000
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "expectedSubmitAt": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "missingReason",
          "ownerId",
          "expectedSubmitAt",
          "version"
        ]
      },
      "MaterialFollowupDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "maxLength": 1000
          },
          "createTask": {
            "type": "object",
            "default": false
          },
          "dueAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "note"
        ]
      },
      "ArchiveMaterialDto": {
        "type": "object",
        "properties": {
          "archiveStatus": {
            "type": "string",
            "enum": [
              "ARCHIVED",
              "ARCHIVE_FAILED"
            ]
          },
          "note": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "archiveStatus",
          "note"
        ]
      },
      "CreateMaterialSubmissionDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "object",
            "maxLength": 1000,
            "nullable": true,
            "description": "管家代传时必填"
          }
        }
      },
...
      "ApiEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data",
          "error",
          "requestId"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "nullable": true
          },
          "error": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              }
            ],
            "nullable": true
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          }
        }
      }
    }
  }
}